data-pipeline/requirements.txt
httpx==0.27.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
python-dotenv==1.0.1
gtfs-kit==12.0.2
pandas==2.2.2
numpy==1.26.4
pydantic==2.7.4
pydantic-settings==2.3.1
tenacity==8.3.0
structlog==24.2.0
